solved bug np

+++ ScintSuite/SimulationCodes/FILDSIM/forwardModelling.py
@@ -576,7 +576,7 @@ def synthetic_signal(pinhole_distribution: dict, efficiency, optics_parameters,
     # Consider the exposure time
     synthetic_frame *= exp_time
     # Pass to integer, as we are dealing with counts
-    synthetic_frame = synthetic_frame.astype(np.int)
+    synthetic_frame = synthetic_frame.astype(int)
     original_frame = synthetic_frame.copy()
     # --- Add noise
     noise = {
@@ -589,7 +589,7 @@ def synthetic_signal(pinhole_distribution: dict, efficiency, optics_parameters,
         'neutrons': None,
         'gamma': None,
         'camera_gamma': None,
-        'total': np.zeros(synthetic_frame.shape, np.int),
+        'total': np.zeros(synthetic_frame.shape, int),
     }
     # broken pixels or fibers
     dummy = ssnoise.broken_pixels(synthetic_frame, **noise_options['broken'])
